+++ tests/Commands/MakeMiddlewareCommandTest.php
@@ -5,7 +5,7 @@ namespace Tests\Commands;
 test('can make middleware', function () {
     $this->artisan('beyond:make:middleware Admin/User/PremiumUserMiddleware');
 
-    expect(base_path() . '/src/App/Admin/User/Middleware/PremiumUserMiddleware.php')
+    expect(base_path() . '/src/App/Admin/User/Middlewares/PremiumUserMiddleware.php')
         ->toBeFile()
         ->toMatchNamespaceAndClassName()
         ->toPlaceholdersBeReplaced();
@@ -14,7 +14,7 @@ test('can make middleware', function () {
 test('can make support middleware', function () {
     $this->artisan('beyond:make:middleware PremiumUserMiddleware --support');
 
-    expect(base_path() . '/src/Support/Packages/Laravel/Middleware/PremiumUserMiddleware.php')
+    expect(base_path() . '/src/Support/Middlewares/PremiumUserMiddleware.php')
         ->toBeFile()
         ->toMatchNamespaceAndClassName()
         ->toPlaceholdersBeReplaced();